+++ src/stores/RnStacker.ts
@@ -2,6 +2,7 @@ import { action, observable } from 'mobx'
 import type { ReactComponentLike } from 'prop-types'
 import type { SyntheticEvent } from 'react'
 
+import { defaultTimeout } from '#/config'
 import { ctx } from '#/stores/ctx'
 import { RnKeyboard } from '#/stores/RnKeyboard'
 import { BackgroundTimer } from '#/utils/BackgroundTimer'
@@ -57,7 +58,7 @@ export class RnStackerStore {
           this.stackAnimating = true
           BackgroundTimer.setTimeout(() => {
             this.stackAnimating = false
-          }, 300)
+          }, defaultTimeout)
         }
         //
         const stack0 = {} as RnStack