+++ src/stores/callStore.ts
@@ -7,7 +7,7 @@ import { v4 as newUuid } from 'uuid'
 
 import { mdiPhone } from '#/assets/icons'
 import type { MakeCallFn, PbxPhoneappliContact, Session } from '#/brekekejs'
-import { isAndroid, isIos, isWeb } from '#/config'
+import { defaultTimeout, isAndroid, isIos, isWeb } from '#/config'
 import { addCallHistory } from '#/stores/addCallHistory'
 import type { ConnectionState } from '#/stores/authStore'
 import { Call } from '#/stores/Call'
@@ -580,10 +580,10 @@ export class CallStore {
   startCall: MakeCallFn = async (number: string, ...args) => {
     // make sure sip is ready before make call
     if (ctx.auth.sipState !== 'success') {
-      return
+      return false
     }
     if (!(await permForCall())) {
-      return
+      return false
     }
     if (
       this.callkeepUuidPending ||
@@ -592,7 +592,7 @@ export class CallStore {
       RnAlert.error({
         message: intlDebug`There is already an outgoing call`,
       })
-      return
+      return false
     }
     // check line resource
     if (
@@ -606,7 +606,7 @@ export class CallStore {
         )
         args[0] = { ...args[0], extraHeaders }
       } catch (err) {
-        return
+        return false
       }
     }
     // start call logic in RNCallKeep
@@ -646,11 +646,11 @@ export class CallStore {
       sipState === 'stopped'
     ) {
       ctx.auth.reconnectAndWaitSip().then(sipCreateSession)
-      return
+      return true
     }
     if (sipState === 'connecting') {
       ctx.auth.waitSip().then(sipCreateSession)
-      return
+      return true
     }
     // in case of sip state is success
     // there could still cases that the sip is disconnected but state not updated yet
@@ -703,6 +703,7 @@ export class CallStore {
       }),
       500,
     )
+    return true
   }
   startVideoCall = (number: string) => this.startCall(number, undefined, true)
 
@@ -724,7 +725,7 @@ export class CallStore {
   }
   private updateBackgroundCallsDebounce = debounce(
     this.updateBackgroundCalls,
-    300,
+    defaultTimeout,
     { maxWait: 1000 },
   )
   @action private updateCurrentCall = () => {
@@ -741,9 +742,13 @@ export class CallStore {
     }
     this.updateBackgroundCallsDebounce()
   }
-  private updateCurrentCallDebounce = debounce(this.updateCurrentCall, 300, {
-    maxWait: 1000,
-  })
+  private updateCurrentCallDebounce = debounce(
+    this.updateCurrentCall,
+    defaultTimeout,
+    {
+      maxWait: 1000,
+    },
+  )
 
   // callkeep + pn data
   @observable callkeepMap: {
@@ -1054,7 +1059,7 @@ export class CallStore {
         if (await BrekekeUtils.isLocked()) {
           BrekekeUtils.setIsAppActive(false, true)
         }
-      }, 300)
+      }, defaultTimeout)
     })
   }